refactor(sync): rename p2p to checkpoint
This is more indicative of its function. It contains checkpoint
sync related code.
